backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
redis==5.0.3
celery==5.3.6
flower==2.0.1
python-dotenv==1.0.1
cryptography==46.0.6
scrapy==2.11.2
fake-useragent==1.5.1
pytest==8.1.1
pytest-django==4.8.0
